+++ src/logic/entities/components/Component.h
@@ -12,7 +12,7 @@ class Component {
 public:
 	virtual ~Component();
 
-	virtual void update(const long millis) = 0;
+	virtual void update(const float millis) = 0;
 };
 
 #endif /* LOGIC_ENTITIES_COMPONENTS_COMPONENT_H_ */
